+++ sdk/storage/azure-storage-blobs/test/block_blob_client_test.cpp
@@ -109,6 +109,30 @@ namespace Azure { namespace Storage { namespace Test {
     EXPECT_EQ(res->BlobSize, static_cast<int64_t>(m_blobContent.size()));
   }
 
+  TEST_F(BlockBlobClientTest, DownloadTransactionalHash)
+  {
+    const int64_t downloadLength = 1024;
+    Blobs::DownloadBlobOptions options;
+    options.Range = Azure::Core::Http::Range();
+    options.Range.GetValue().Offset = 0;
+    options.Range.GetValue().Length = downloadLength;
+    options.RangeHashAlgorithm = HashAlgorithm::Md5;
+    auto res = m_blockBlobClient->Download(options);
+    ASSERT_TRUE(res->TransactionalContentHash.HasValue());
+    EXPECT_EQ(res->TransactionalContentHash.GetValue().Algorithm, HashAlgorithm::Md5);
+    EXPECT_EQ(
+        res->TransactionalContentHash.GetValue().Value,
+        Md5::Hash(m_blobContent.data(), downloadLength));
+
+    options.RangeHashAlgorithm = HashAlgorithm::Crc64;
+    res = m_blockBlobClient->Download(options);
+    ASSERT_TRUE(res->TransactionalContentHash.HasValue());
+    EXPECT_EQ(res->TransactionalContentHash.GetValue().Algorithm, HashAlgorithm::Crc64);
+    EXPECT_EQ(
+        res->TransactionalContentHash.GetValue().Value,
+        Crc64::Hash(m_blobContent.data(), downloadLength));
+  }
+
   TEST_F(BlockBlobClientTest, DISABLED_LastAccessTime)
   {
     {
@@ -351,7 +375,7 @@ namespace Azure { namespace Storage { namespace Test {
     EXPECT_TRUE(IsValidTime(res->LastModified));
     EXPECT_TRUE(IsValidTime(res->CreatedOn));
     EXPECT_EQ(res->Metadata, m_blobUploadOptions.Metadata);
-    EXPECT_EQ(res->ContentLength, static_cast<int64_t>(m_blobContent.size()));
+    EXPECT_EQ(res->BlobSize, static_cast<int64_t>(m_blobContent.size()));
     EXPECT_EQ(res->HttpHeaders, m_blobUploadOptions.HttpHeaders);
     EXPECT_EQ(res->HttpHeaders.ContentHash.Algorithm, Storage::HashAlgorithm::Md5);
     EXPECT_EQ(res->Tier.GetValue(), Azure::Storage::Blobs::Models::AccessTier::Cool);
@@ -385,7 +409,7 @@ namespace Azure { namespace Storage { namespace Test {
     EXPECT_FALSE(res.GetRawResponse().GetHeaders().at(Details::HttpHeaderXMsVersion).empty());
     EXPECT_TRUE(res->ETag.HasValue());
     EXPECT_TRUE(IsValidTime(res->LastModified));
-    EXPECT_EQ(res->ContentLength, static_cast<int64_t>(block1Content.size()));
+    EXPECT_EQ(res->BlobSize, static_cast<int64_t>(block1Content.size()));
     ASSERT_FALSE(res->CommittedBlocks.empty());
     EXPECT_EQ(res->CommittedBlocks[0].Name, blockId1);
     EXPECT_EQ(res->CommittedBlocks[0].Size, static_cast<int64_t>(block1Content.size()));
@@ -395,15 +419,14 @@ namespace Azure { namespace Storage { namespace Test {
     Blobs::GetBlockListOptions options2;
     options2.ListType = Blobs::Models::BlockListTypeOption::All;
     res = blockBlobClient.GetBlockList(options2);
-    EXPECT_EQ(res->ContentLength, static_cast<int64_t>(block1Content.size()));
+    EXPECT_EQ(res->BlobSize, static_cast<int64_t>(block1Content.size()));
     ASSERT_FALSE(res->UncommittedBlocks.empty());
     EXPECT_EQ(res->UncommittedBlocks[0].Name, blockId2);
     EXPECT_EQ(res->UncommittedBlocks[0].Size, static_cast<int64_t>(m_blobContent.size()));
 
     blockBlobClient.CommitBlockList({blockId1, blockId2});
     res = blockBlobClient.GetBlockList(options2);
-    EXPECT_EQ(
-        res->ContentLength, static_cast<int64_t>(block1Content.size() + m_blobContent.size()));
+    EXPECT_EQ(res->BlobSize, static_cast<int64_t>(block1Content.size() + m_blobContent.size()));
     EXPECT_TRUE(res->UncommittedBlocks.empty());
   }
 
@@ -469,8 +492,10 @@ namespace Azure { namespace Storage { namespace Test {
       {
         auto res
             = m_blockBlobClient->DownloadTo(downloadBuffer.data(), downloadBuffer.size(), options);
-        EXPECT_EQ(res->ContentLength, actualDownloadSize);
-        downloadBuffer.resize(static_cast<std::size_t>(res->ContentLength));
+        EXPECT_EQ(res->BlobSize, blobSize);
+        EXPECT_EQ(res->ContentRange.Length.GetValue(), actualDownloadSize);
+        EXPECT_EQ(res->ContentRange.Offset, offset.HasValue() ? offset.GetValue() : 0);
+        downloadBuffer.resize(static_cast<std::size_t>(res->ContentRange.Length.GetValue()));
         EXPECT_EQ(downloadBuffer, expectedData);
       }
       else
@@ -538,7 +563,9 @@ namespace Azure { namespace Storage { namespace Test {
       if (actualDownloadSize > 0)
       {
         auto res = m_blockBlobClient->DownloadTo(tempFilename, options);
-        EXPECT_EQ(res->ContentLength, actualDownloadSize);
+        EXPECT_EQ(res->BlobSize, blobSize);
+        EXPECT_EQ(res->ContentRange.Length.GetValue(), actualDownloadSize);
+        EXPECT_EQ(res->ContentRange.Offset, offset.HasValue() ? offset.GetValue() : 0);
         EXPECT_EQ(ReadFile(tempFilename), expectedData);
       }
       else
@@ -618,6 +645,20 @@ namespace Azure { namespace Storage { namespace Test {
                 downloadBuffer.data(), static_cast<std::size_t>(length - 1), options),
             std::runtime_error);
       }
+
+      // initial chunk size
+      futures.emplace_back(
+          std::async(std::launch::async, testDownloadToBuffer, c, blobSize, 0, 1024, 512, 1024));
+      futures.emplace_back(
+          std::async(std::launch::async, testDownloadToFile, c, blobSize, 0, 1024, 512, 1024));
+      futures.emplace_back(
+          std::async(std::launch::async, testDownloadToBuffer, c, blobSize, 0, 1024, 1024, 1024));
+      futures.emplace_back(
+          std::async(std::launch::async, testDownloadToFile, c, blobSize, 0, 1024, 1024, 1024));
+      futures.emplace_back(
+          std::async(std::launch::async, testDownloadToBuffer, c, blobSize, 0, 1024, 2048, 1024));
+      futures.emplace_back(
+          std::async(std::launch::async, testDownloadToFile, c, blobSize, 0, 1024, 2048, 1024));
     }
     for (auto& f : futures)
     {
@@ -680,14 +721,16 @@ namespace Azure { namespace Storage { namespace Test {
     blockBlobClient.SetMetadata(m_blobUploadOptions.Metadata);
 
     auto res = blockBlobClient.DownloadTo(emptyContent.data(), 0);
-    EXPECT_EQ(res->ContentLength, 0);
+    EXPECT_EQ(res->BlobSize, 0);
+    EXPECT_EQ(res->ContentRange.Length.GetValue(), 0);
     EXPECT_TRUE(res->ETag.HasValue());
     EXPECT_TRUE(IsValidTime(res->LastModified));
     EXPECT_EQ(res->HttpHeaders, m_blobUploadOptions.HttpHeaders);
     EXPECT_EQ(res->Metadata, m_blobUploadOptions.Metadata);
     EXPECT_EQ(res->BlobType, Azure::Storage::Blobs::Models::BlobType::BlockBlob);
     res = blockBlobClient.DownloadTo(tempFilename);
-    EXPECT_EQ(res->ContentLength, 0);
+    EXPECT_EQ(res->BlobSize, 0);
+    EXPECT_EQ(res->ContentRange.Length.GetValue(), 0);
     EXPECT_TRUE(res->ETag.HasValue());
     EXPECT_TRUE(IsValidTime(res->LastModified));
     EXPECT_EQ(res->HttpHeaders, m_blobUploadOptions.HttpHeaders);
@@ -697,14 +740,16 @@ namespace Azure { namespace Storage { namespace Test {
     DeleteFile(tempFilename);
 
     res = blockBlobClient.DownloadTo(emptyContent.data(), static_cast<std::size_t>(8_MB));
-    EXPECT_EQ(res->ContentLength, 0);
+    EXPECT_EQ(res->BlobSize, 0);
+    EXPECT_EQ(res->ContentRange.Length.GetValue(), 0);
     EXPECT_TRUE(res->ETag.HasValue());
     EXPECT_TRUE(IsValidTime(res->LastModified));
     EXPECT_EQ(res->HttpHeaders, m_blobUploadOptions.HttpHeaders);
     EXPECT_EQ(res->Metadata, m_blobUploadOptions.Metadata);
     EXPECT_EQ(res->BlobType, Azure::Storage::Blobs::Models::BlobType::BlockBlob);
     res = blockBlobClient.DownloadTo(tempFilename);
-    EXPECT_EQ(res->ContentLength, 0);
+    EXPECT_EQ(res->BlobSize, 0);
+    EXPECT_EQ(res->ContentRange.Length.GetValue(), 0);
     EXPECT_TRUE(res->ETag.HasValue());
     EXPECT_TRUE(IsValidTime(res->LastModified));
     EXPECT_EQ(res->HttpHeaders, m_blobUploadOptions.HttpHeaders);
@@ -722,14 +767,16 @@ namespace Azure { namespace Storage { namespace Test {
 
       res = blockBlobClient.DownloadTo(
           emptyContent.data(), static_cast<std::size_t>(8_MB), options);
-      EXPECT_EQ(res->ContentLength, 0);
+      EXPECT_EQ(res->BlobSize, 0);
+      EXPECT_EQ(res->ContentRange.Length.GetValue(), 0);
       EXPECT_TRUE(res->ETag.HasValue());
       EXPECT_TRUE(IsValidTime(res->LastModified));
       EXPECT_EQ(res->HttpHeaders, m_blobUploadOptions.HttpHeaders);
       EXPECT_EQ(res->Metadata, m_blobUploadOptions.Metadata);
       EXPECT_EQ(res->BlobType, Azure::Storage::Blobs::Models::BlobType::BlockBlob);
       res = blockBlobClient.DownloadTo(tempFilename, options);
-      EXPECT_EQ(res->ContentLength, 0);
+      EXPECT_EQ(res->BlobSize, 0);
+      EXPECT_EQ(res->ContentRange.Length.GetValue(), 0);
       EXPECT_TRUE(res->ETag.HasValue());
       EXPECT_TRUE(IsValidTime(res->LastModified));
       EXPECT_EQ(res->HttpHeaders, m_blobUploadOptions.HttpHeaders);
@@ -788,7 +835,7 @@ namespace Azure { namespace Storage { namespace Test {
       EXPECT_TRUE(IsValidTime(res->LastModified));
       auto properties = *blockBlobClient.GetProperties();
       properties.HttpHeaders.ContentHash.Value.clear();
-      EXPECT_EQ(properties.ContentLength, blobSize);
+      EXPECT_EQ(properties.BlobSize, blobSize);
       EXPECT_EQ(properties.HttpHeaders, options.HttpHeaders);
       EXPECT_EQ(properties.Metadata, options.Metadata);
       EXPECT_EQ(properties.Tier.GetValue(), options.Tier.GetValue());
@@ -823,7 +870,7 @@ namespace Azure { namespace Storage { namespace Test {
       EXPECT_TRUE(IsValidTime(res->LastModified));
       auto properties = *blockBlobClient.GetProperties();
       properties.HttpHeaders.ContentHash.Value.clear();
-      EXPECT_EQ(properties.ContentLength, blobSize);
+      EXPECT_EQ(properties.BlobSize, blobSize);
       EXPECT_EQ(properties.HttpHeaders, options.HttpHeaders);
       EXPECT_EQ(properties.Metadata, options.Metadata);
       EXPECT_EQ(properties.Tier.GetValue(), options.Tier.GetValue());